requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.32.5
gunicorn==21.2.0
pydantic==2.5.0
bcrypt==4.1.2